Export needed typings in web3-eth2-core. Export class definition in web3-eth2-beaconchain
