+++ packages/web3-eth2-beaconchain/types/index.d.ts
@@ -17,6 +17,8 @@ import {
   SignedVoluntaryExit
 } from '@chainsafe/lodestar-types'
 
+import { IBaseAPISchema, ETH2BaseOpts } from 'web3-eth2-core'
+
 export type StateId = 'head' | 'genesis' | 'finalized' | 'justified' | Slot | Root
 export type BlockId = 'head' | 'genesis' | 'finalized' | Slot | Root
 
@@ -43,4 +45,13 @@ export interface IETH2BeaconChain {
   submitProposerSlashings(): Promise<void>
   getSignedVoluntaryExits(): Promise<SignedVoluntaryExit[]>
   submitVoluntaryExit(): Promise<void>
+}
+
+// @ts-ignore
+export class ETH2BeaconChain implements IETH2BeaconChain {
+  constructor(
+    provider: string,
+    schema: IBaseAPISchema,
+    opts: ETH2BaseOpts
+  )
 }
