+++ packages/web3-eth2-core/types/index.d.ts
@@ -3,3 +3,20 @@ export interface ETH2BaseOpts {
 }
 
 export type ETH2Function<T = any> = (...args: any[]) => Promise<T>;
+
+export interface IBaseAPISchema {
+    packageName: string,
+    routePrefix: string,
+    methods: IBaseAPIMethodSchema[]
+}
+
+export interface IBaseAPIMethodSchema {
+    notImplemented?: true,
+    name: string,
+    route: string,
+    restMethod: string
+    inputFormatter: any,
+    outputFormatter: any,
+    errors: any
+    errorPrefix: string
+}
